Add UVSphere
Rename Sphere to QuadSphere.

Apply style fixes.

+++ first-simple-procedural-geometry/Assets/Square.cs
@@ -24,10 +24,10 @@ public class Square : MonoBehaviour {
         // };
 
         // Level-Of-Details (count of quads along one side)
-        float xyQuadSize = (float)SIZE / LOD; // discrete quad's side xy size (optimization)
-        float uvQuadSize = 1f / LOD;          // discrete quad's side uv size (optimization)
-        float halfSize   = SIZE / 2f;         // hald of the SIZE (optimization)
-        int quadLOD      = (int)(LOD * LOD);  // LOD^2 (optimization)
+        float xyQuadSize = (float)SIZE / (int)LOD; // discrete quad's side xy size
+        float uvQuadSize = 1f / (int)LOD;          // discrete quad's side uv size
+        float halfSize   = SIZE / 2f;              // half of the SIZE
+        int quadLOD      = (int)(LOD * LOD);       // LOD^2
 
         Vector3[] vertices = new Vector3[4 * quadLOD]; // 4*LOD^2 vertices
         Vector3[] normals  = new Vector3[4 * quadLOD]; // 4*LOD^2 normals
@@ -37,22 +37,22 @@ public class Square : MonoBehaviour {
         for (int i = 0; i < quadLOD; ++i) {
             int row    = i / (int)LOD;
             int column = i % (int)LOD;
-            float lux  = column * xyQuadSize - halfSize;
-            float rux  = (column + 1) * xyQuadSize - halfSize;
-            float ldx  = lux;
-            float rdx  = rux;
-            float luy  = (row + 1) * xyQuadSize - halfSize;
-            float ruy  = luy;
-            float ldy  = row * xyQuadSize - halfSize;
-            float rdy  = ldy;
-            float luu  = column * uvQuadSize;
-            float ruu  = (column + 1) * uvQuadSize;
-            float ldu  = luu;
-            float rdu  = ruu;
-            float luv  = (row + 1) * uvQuadSize;
-            float ruv  = luv;
-            float ldv  = row * uvQuadSize;
-            float rdv  = ldv;
+            float lux = column * xyQuadSize - halfSize;
+            float rux = (column + 1) * xyQuadSize - halfSize;
+            float ldx = lux;
+            float rdx = rux;
+            float luy = (row + 1) * xyQuadSize - halfSize;
+            float ruy = luy;
+            float ldy = row * xyQuadSize - halfSize;
+            float rdy = ldy;
+            float luu = column * uvQuadSize;
+            float ruu = (column + 1) * uvQuadSize;
+            float ldu = luu;
+            float rdu = ruu;
+            float luv = (row + 1) * uvQuadSize;
+            float ruv = luv;
+            float ldv = row * uvQuadSize;
+            float rdv = ldv;
 
             vertices[i * 4]     = new Vector3(lux, luy, 0); // left-up
             vertices[i * 4 + 1] = new Vector3(rux, ruy, 0); // right-up
@@ -69,12 +69,12 @@ public class Square : MonoBehaviour {
             uvs[i * 4 + 2] = new Vector2(ldu, ldv); // left-down
             uvs[i * 4 + 3] = new Vector2(rdu, rdv); // right-down
 
-            triangles[i * 6]     = i * 4;     // left-up quad
-            triangles[i * 6 + 1] = i * 4 + 1; // left-up quad
-            triangles[i * 6 + 2] = i * 4 + 2; // left-up quad
-            triangles[i * 6 + 3] = i * 4 + 1; // right-down quad
-            triangles[i * 6 + 4] = i * 4 + 3; // right-down quad
-            triangles[i * 6 + 5] = i * 4 + 2; // right-down quad
+            triangles[i * 6]     = i * 4;     // left-up triangle
+            triangles[i * 6 + 1] = i * 4 + 1; // left-up triangle
+            triangles[i * 6 + 2] = i * 4 + 2; // left-up triangle
+            triangles[i * 6 + 3] = i * 4 + 1; // right-down triangle
+            triangles[i * 6 + 4] = i * 4 + 3; // right-down triangle
+            triangles[i * 6 + 5] = i * 4 + 2; // right-down triangle
         }
 
         mesh.vertices  = vertices;
